Handle identical x-values in apply_curve
When the lowest and highest x-value in curve are identical, apply_curve
would choose to interpolate based on the lowest value. Instead, we want
to use the average of the lowest and highest. This has mostly an effect
in variables such as precipitation, where there are many 0-values in the
curve.