Fixed make install-python-user and install-python
Didn't work on python3.10 since it couldn't find gridpp_swig.so. This is
because CMake generates _gridpp_swig.so, and the either the wrong name
was hardcoded in CMakeFiles, or CMake has changed behaviour and calls
it by the new name.

Turns out that removing the custom installation class in setup.py still
works. Make sense, since install-develop-user and install-develop does
not use the custom installation class.