Use 3D coordinates when computing distance in structure function
Previously, a crude approximation based on lat/lon was used in the
structure function to save time. In this commit, we use the 3D
coordinates used by KDTree to compute a straight-line distance. This is
even faster than the crude method and more robust in the poles. It is
also more accurate for computing the straight-line distance.

This requires the KDTree to keep an extra copy of the x,y,z coordinates,
which may increase the memory footprint.

This commit improves the computational performance of
optimal_interpolation by 23%, according to benchmark.py.